@@ -39,13 +39,13 @@
     "magento/module-sales": ">=101",
     "nyholm/psr7": "^1.2",
     "pay-now/paynow-php-sdk": "^2.3.1",
-    "php": ">=7.1",
+    "php": ">=7.2",
     "php-http/curl-client": "^2.1",
     "ext-json": "*"
   },
   "require-dev": {
     "magento/magento-coding-standard": "2.*",
-    "phpunit/phpunit": "~6.2.0",
+    "phpunit/phpunit": "^8.5.36",
     "squizlabs/php_codesniffer": "^3.6"
   },
   "autoload": {
