@@ -126,7 +126,7 @@ class Success extends Action
             $loggerContext['exception'] = [
                 'message' => $exception->getMessage(),
                 'file' => $exception->getFile(),
-                'line' => $exception->line(),
+                'line' => $exception->getLine(),
             ];
             $this->logger->error(
                 'Error occurred handling notification',
